refactored search page view into multiple components
