+++ hosting/src/app/pages/search/header/header.component.scss
@@ -1,0 +1,34 @@
+header {
+  margin-top: 10vh;
+  margin-bottom: 1rem;
+
+  h1 {
+    margin: 0 0 1rem 0;
+  }
+
+  & > div {
+    display: flex;
+    & > :nth-child(1) {
+      flex: 4;
+    }
+    & > :nth-child(2) {
+      flex: 1;
+    }
+
+    .proposals {
+      margin-bottom: 1rem;
+      & > span {
+        font-size: 0.7rem;
+        opacity: 0.6;
+      }
+    }
+    .checkboxes mat-checkbox {
+      margin-right: 2rem;
+    }
+
+    mat-form-field {
+      width: 100%;
+    }
+  }
+}
+
