+++ hosting/src/app/pages/search/hints/hints.component.scss
@@ -1,0 +1,13 @@
+p {
+  max-width: 40rem;
+}
+
+h2 {
+  display: flex;
+  align-items: center;
+  mat-icon {
+    margin-right: 1rem;
+  }
+}
+
+
