+++ hosting/src/app/pages/search/search.component.scss
@@ -12,69 +12,3 @@
   }
 }
 
-p {
-  max-width: 40rem;
-}
-
-h2 {
-  display: flex;
-  align-items: center;
-  mat-icon {
-    margin-right: 1rem;
-  }
-}
-
-header {
-  margin-top: 10vh;
-  margin-bottom: 1rem;
-
-  h1 {
-    margin: 0 0 1rem 0;
-  }
-
-  & > div {
-    display: flex;
-    & > :nth-child(1) {
-      flex: 4;
-    }
-    & > :nth-child(2) {
-      flex: 1;
-    }
-
-    .proposals {
-      margin-bottom: 1rem;
-      & > span {
-        font-size: 0.7rem;
-        opacity: 0.6;
-      }
-    }
-    .checkboxes mat-checkbox {
-      margin-right: 2rem;
-    }
-
-    mat-form-field {
-      width: 100%;
-    }
-  }
-}
-
-main {
-  display: flex;
-  flex-direction: column;
-
-  &.single {
-    flex-direction: row;
-    flex-wrap: wrap;
-    justify-content: space-between;
-  }
-
-  section {
-    & > div {
-      display: flex;
-      margin: 0 -1rem;
-      justify-content: space-between;
-      align-items: center;
-      overflow: hidden;
-    }
-  }
-}
