+++ hosting/src/app/pages/search/single-type/single-type.component.html
@@ -1,0 +1,10 @@
+<app-article-card *ngFor="let article of articles$ | async" [data]="article"></app-article-card>
+<app-image-card *ngFor="let image of images$ | async" [data]="image"></app-image-card>
+<app-event-card *ngFor="let event of events$ | async" [data]="event"></app-event-card>
+<app-paginator
+  *ngIf="(searchStrings$ | async)?.length === 0"
+  class="mat-elevation-z2"
+  [pageIndex]="pageIndex$ | async"
+  [pageSize]="limit"
+  [length]="length$ | async"
+></app-paginator>
