+++ hosting/src/app/pages/search/single-type/single-type.component.scss
@@ -1,0 +1,8 @@
+:host {
+  display: flex;
+  flex-direction: column;
+  flex-direction: row;
+  flex-wrap: wrap;
+  justify-content: space-between;
+  margin: 0 -1rem;
+}
